requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==3.0.0
reportlab==4.4.9
numpy==2.4.1
python-dateutil==2.9.0.post0
Pillow==12.1.0
